requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
environ==1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
